# cashmanager/src/main/java/eu/epitech/cashmanager/domain/Product.java
package eu.epitech.cashmanager.domain;

import com.fasterxml.jackson.annotation.JsonIgnore;

import javax.persistence.*;
import java.util.List;
import java.util.Set;
import java.util.UUID;

@Entity
public class Product {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private int id;
    private float price;
    private String name;
    private String description;

    @Column(unique = true)
    private UUID uuid;

    @OneToMany(mappedBy = "product")
    @JsonIgnore
    private Set<BillContent> billContents;

    @OneToMany(fetch = FetchType.LAZY, cascade = CascadeType.REMOVE, mappedBy = "product")
    private List<ProductPicture> pictures;

    public Product() {
        this.uuid = UUID.randomUUID();
    }

    public Product(int id, String name, String description, float price, List<ProductPicture> pictures) {
        super();
        this.id = id;
        this.name = name;
        this.description = description;
        this.price = price;
        this.pictures = pictures;
        this.uuid = UUID.randomUUID();
    }

    public int getId() {
        return id;
    }

    public float getPrice() {
        return price;
    }

    public void setPrice(float price) {
        this.price = price;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public UUID getUuid() {
        return uuid;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public Set<BillContent> getBillContents() {
        return billContents;
    }

    public List<ProductPicture> getPictures() {
        return pictures;
    }

    public void setPictures(List<ProductPicture> pictures) {
        this.pictures = pictures;
    }
}